@@ -1,5 +1,5 @@
 [package]
-edition = "2018"
+edition = "2021"
 license = "MIT OR Apache-2.0"
 name = "hotham-stress-test"
 version = "0.2.0"
@@ -12,7 +12,7 @@ name = "hotham_stress_test_desktop"
 path = "src/main.rs"
 
 [dependencies]
-hotham = {path = "../../hotham"}
+hotham = { path = "../../hotham" }
 
 [target.'cfg(target_os = "android")'.dependencies]
 ndk-glue = "0.6"
@@ -49,7 +49,10 @@ value = "vr_only"
 
 [[package.metadata.android.application.activity.intent_filter]]
 actions = ["android.intent.action.MAIN"]
-categories = ["com.oculus.intent.category.VR", "android.intent.category.LAUNCHER"]
+categories = [
+    "com.oculus.intent.category.VR",
+    "android.intent.category.LAUNCHER",
+]
 
 [[package.metadata.android.application.activity.meta_data]]
 name = "com.oculus.vr.focusaware"
