@@ -1,8 +1,8 @@
 [package]
-edition = "2018"
+edition = "2021"
 license = "MIT OR Apache-2.0"
 name = "hotham-examples"
 version = "0.2.0"
 
 [dependencies]
-hotham = {path = "../../hotham"}
+hotham = { path = "../../hotham" }
