@@ -1,21 +1,17 @@
 [package]
-edition = "2018"
+edition = "2021"
 license = "MIT OR Apache-2.0"
 name = "simple-scene-example"
 version = "0.2.0"
 
 [lib]
-crate-type = ["lib", "cdylib"]
-
-[[bin]]
-name = "hotham_simple_scene_example"
-path = "src/main.rs"
+crate-type = ["rlib", "cdylib"]
 
 [dependencies]
-hotham = {path = "../../hotham"}
+hotham = { path = "../../hotham" }
 
 [target.'cfg(target_os = "android")'.dependencies]
-ndk-glue = "0.6"
+ndk-glue = "0.7"
 
 [package.metadata.android]
 apk_label = "Hotham Simple Scene Example"
@@ -49,7 +45,10 @@ value = "vr_only"
 
 [[package.metadata.android.application.activity.intent_filter]]
 actions = ["android.intent.action.MAIN"]
-categories = ["com.oculus.intent.category.VR", "android.intent.category.LAUNCHER"]
+categories = [
+    "com.oculus.intent.category.VR",
+    "android.intent.category.LAUNCHER",
+]
 
 [[package.metadata.android.application.activity.meta_data]]
 name = "com.oculus.vr.focusaware"
