@@ -1,6 +1,3 @@
-use crab_saber::real_main;
-use hotham::HothamResult;
-
-fn main() -> HothamResult<()> {
-    real_main()
+fn main() {
+    crab_saber::real_main().unwrap()
 }
