@@ -38,7 +38,7 @@ impl<T: Sized> Buffer<T> {
 
         let buffer = device
             .create_buffer(
-                &vk::BufferCreateInfo::builder().usage(usage).size(size),
+                &vk::BufferCreateInfo::default().usage(usage).size(size),
                 None,
             )
             .unwrap();
@@ -60,13 +60,14 @@ impl<T: Sized> Buffer<T> {
             )
             .unwrap();
 
-        // Transmute the pointer into GPU memory so that we can easily access it again.
-        let memory_address = std::mem::transmute(memory_address);
+        // Stash the pointer so that we can easily access it again.
+        let memory_address = std::mem::transmute::<*mut std::ffi::c_void, *mut T>(memory_address);
+        let memory_address = std::ptr::NonNull::new_unchecked(memory_address);
 
         Self {
             buffer,
             device_memory,
-            memory_address: std::ptr::NonNull::new_unchecked(memory_address),
+            memory_address,
             len: 0,
             max_len,
             usage,
@@ -144,7 +145,7 @@ impl<T: Sized> Buffer<T> {
         descriptor_set: vk::DescriptorSet,
         binding: u32,
     ) {
-        let buffer_info = vk::DescriptorBufferInfo::builder()
+        let buffer_info = vk::DescriptorBufferInfo::default()
             .buffer(self.buffer)
             .offset(0)
             .range(vk::WHOLE_SIZE);
@@ -155,7 +156,7 @@ impl<T: Sized> Buffer<T> {
             vk::DescriptorType::STORAGE_BUFFER
         };
 
-        let write = vk::WriteDescriptorSet::builder()
+        let write = vk::WriteDescriptorSet::default()
             .buffer_info(std::slice::from_ref(&buffer_info))
             .dst_set(descriptor_set)
             .dst_binding(binding)
